requirements.txt
# Google ADK 기반 에이전트 시스템 - 프로덕션 의존성

# Google ADK 및 AI 모델
google-adk>=0.1.0
google-cloud-aiplatform>=1.40.0
google-generativeai>=0.3.0
vertexai>=1.40.0

# 웹 프레임워크
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 벡터 데이터베이스
qdrant-client>=1.7.0
chromadb>=0.4.0

# LLM 및 AI 도구
langchain>=0.1.0
langchain-google-genai>=0.0.8
langchain-community>=0.0.10
litellm>=1.20.0

# 데이터베이스
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0

# HTTP 클라이언트
httpx>=0.25.0
aiohttp>=3.9.0

# 환경 변수 및 설정
python-dotenv>=1.0.0
python-multipart>=0.0.6

# 로깅 및 모니터링
structlog>=23.2.0
rich>=13.7.0

# 유틸리티
pydantic-core>=2.14.0
typing-extensions>=4.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 멀티모달 지원
speech-recognition>=3.10.0
pyttsx3>=2.90

# 데이터 처리
pandas>=2.1.0
numpy>=1.24.0